backend/services/langgraph/requirements.txt
# Core LangGraph & LangChain
langgraph>=0.2.0
langgraph-checkpoint-postgres>=1.0.0
langchain>=0.2.0
langchain-google-genai>=1.0.0
langchain-core>=0.2.0

# API & Web
fastapi>=0.100.0
uvicorn>=0.20.0

# API Clients - Compatible versions
httpx>=0.24.0
twilio>=8.0.0
python-telegram-bot>=20.0

# AI/ML
google-generativeai>=0.3.0
transformers>=4.30.0
torch>=2.0.0
scikit-learn>=1.3.0
numpy>=1.24.0
psutil>=5.9.0

# Database
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0

# Utilities - Let pip resolve compatible versions
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
PyJWT>=2.8.0